- Put in arbitrary day for the datetime formatting - Moved settings.py, cleaned up some stuff
